PR cleanup and edit
