+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/QuotedMessage.kt
@@ -33,7 +33,6 @@ import com.getstream.sdk.chat.utils.extensions.isMine
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.initials
 import io.getstream.chat.android.compose.ui.attachments.content.MessageAttachmentsContent
-import io.getstream.chat.android.compose.ui.attachments.content.QuotedMessageAttachmentContent
 import io.getstream.chat.android.compose.ui.components.avatar.Avatar
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 
@@ -47,7 +46,7 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
  */
 @Deprecated(
     message = "Deprecated in favor of new QuotedMessage implementation that expands on the current one and" +
-        "changes how the quoted message is laid out.",
+        "changes how the quoted message is displayed.",
     replaceWith = ReplaceWith(
         expression =
         "QuotedMessage(message: Message," +
@@ -99,13 +98,12 @@ public fun QuotedMessage(
 }
 
 /**
- * Wraps the quoted message into a special component, open to customization. By default wraps the quoted message
- * into a special component that shows only sender avatar, text and single attachment preview.
+ * Wraps the quoted message into a component that shows only the sender avatar, text and single attachment preview.
  *
  * @param message Message to show.
- * @param modifier Modifier for styling.
  * @param onLongItemClick Handler when the item is long clicked.
  * @param onQuotedMessageClick Handler for quoted message click action.
+ * @param modifier Modifier for styling.
  * @param leadingContent The content shown at the start of the quoted message. By default we provide
  * [DefaultQuotedMessageLeadingContent] which shows the sender avatar in case the sender is not the current user.
  * @param centerContent The content shown at the center of the quoted message. By default we provide
@@ -117,9 +115,9 @@ public fun QuotedMessage(
 @Composable
 public fun QuotedMessage(
     message: Message,
-    modifier: Modifier = Modifier,
     onLongItemClick: (Message) -> Unit,
     onQuotedMessageClick: (Message) -> Unit,
+    modifier: Modifier = Modifier,
     leadingContent: @Composable (Message) -> Unit = { DefaultQuotedMessageLeadingContent(message = it) },
     centerContent: @Composable RowScope.(Message) -> Unit = { DefaultQuotedMessageCenterContent(it) },
     trailingContent: @Composable (Message) -> Unit = { DefaultQuotedMessageTrailingContent(message = it) },
@@ -190,66 +188,14 @@ internal fun DefaultQuotedMessageTrailingContent(message: Message) {
 /**
  * Represents the default content shown in the center of the quoted message wrapped inside a message bubble.
  *
- * By default shows a single message attachment if one is found inside attachments and the message text or the
- * attachment name if no text has been sent.
- *
  * @param message The quoted message.
- * @param attachmentContent The content shown at the start of center content used to preview the attachment. By default
- * we provide [DefaultQuotedMessageAttachmentContent] which show a single attachment preview.
- * @param textContent The content shown at the end of center quoted message content to shows the sent message text or
- * attachment name. By default we provide [DefaultQuotedMessageTextContent] which show the message text if there is any
- * or previewed attachment name.
  */
 @Composable
 public fun RowScope.DefaultQuotedMessageCenterContent(
     message: Message,
-    attachmentContent: @Composable (Message) -> Unit = { DefaultQuotedMessageAttachmentContent(it) },
-    textContent: @Composable (Message) -> Unit = { DefaultQuotedMessageTextContent(it) },
 ) {
-    val isMyMessage = message.isMine()
-
-    val messageBubbleShape = if (isMyMessage) ChatTheme.shapes.myMessageBubble else ChatTheme.shapes.otherMessageBubble
-
-    MessageBubble(
-        modifier = Modifier.weight(1f, fill = false),
-        shape = messageBubbleShape, color = ChatTheme.colors.barsBackground,
-        content = {
-            Row {
-                attachmentContent(message)
-
-                textContent(message)
-            }
-        }
-    )
-}
-
-/**
- * Represents the default attachment preview of the quoted message.
- *
- * By default we show the first attachment that is sent inside the message.
- *
- * @param message The quoted message.
- */
-@Composable
-internal fun DefaultQuotedMessageAttachmentContent(message: Message) {
-    if (message.attachments.isNotEmpty()) {
-        QuotedMessageAttachmentContent(
-            message = message,
-            onLongItemClick = {},
-        )
-    }
-}
-
-/**
- * Represents the default text preview of the quoted message.
- *
- * By default we show the message text if there is any or show the previewed attachment name.
- *
- * @param message The quoted message.
- */
-@Composable
-internal fun DefaultQuotedMessageTextContent(message: Message) {
-    QuotedMessageText(
-        message = message
+    QuotedMessageContent(
+        message = message,
+        modifier = Modifier.weight(1f, fill = false)
     )
 }
