+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/TextUtils.kt
@@ -1,0 +1,100 @@
+/*
+ * Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.chat.android.compose.ui.util
+
+import android.annotation.SuppressLint
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.core.util.PatternsCompat
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.compose.ui.theme.ChatTheme
+
+/**
+ * Takes the given message and builds an annotated message text that shows links and allows for clicks,
+ * if there are any links available.
+ *
+ * @param message The message to extract the text from and style.
+ *
+ * @return The annotated String, with clickable links, if applicable.
+ */
+@Composable
+internal fun buildAnnotatedMessageText(message: Message): AnnotatedString {
+    return buildAnnotatedMessageText(message.text)
+}
+
+/**
+ * Takes the given message text and builds an annotated message text that shows links and allows for clicks,
+ * if there are any links available.
+ *
+ * @param text The message text to style.
+ *
+ * @return The annotated String, with clickable links, if applicable.
+ */
+@Composable
+internal fun buildAnnotatedMessageText(text: String): AnnotatedString {
+    return buildAnnotatedString {
+        // First we add the whole text to the [AnnotatedString] and style it as a regular text.
+        append(text)
+        addStyle(
+            SpanStyle(
+                fontStyle = ChatTheme.typography.body.fontStyle,
+                color = ChatTheme.colors.textHighEmphasis
+            ),
+            start = 0,
+            end = text.length
+        )
+
+        // Then for each available link in the text, we add a different style, to represent the links,
+        // as well as add a String annotation to it. This gives us the ability to open the URL on click.
+        @SuppressLint("RestrictedApi")
+        val matcher = PatternsCompat.AUTOLINK_WEB_URL.matcher(text)
+        while (matcher.find()) {
+            val start = matcher.start()
+            val end = matcher.end()
+
+            addStyle(
+                style = SpanStyle(
+                    color = ChatTheme.colors.primaryAccent,
+                    textDecoration = TextDecoration.Underline,
+                ),
+                start = start,
+                end = end,
+            )
+
+            val linkText = requireNotNull(matcher.group(0)!!)
+
+            // Add "http://" prefix if link has no scheme in it
+            val url = if (URL_SCHEMES.none { scheme -> linkText.startsWith(scheme) }) {
+                URL_SCHEMES[0] + linkText
+            } else {
+                linkText
+            }
+
+            addStringAnnotation(
+                tag = "URL",
+                annotation = url,
+                start = start,
+                end = end,
+            )
+        }
+    }
+}
+
+private val URL_SCHEMES = listOf("http://", "https://")
